+++ src/main/java/com/example/project/service/category/CategoryService.java
@@ -35,17 +35,20 @@ public class CategoryService implements ICategoryService {
 
     @Override
     public CategoryDTO findByName(String name) {
+        if (name == null) return null;
         return convertToDTO(categoryRepository.findByName(name));
     }
 
     @Override
     public CategoryDTO findById(Long id) {
+        if (id == null) return null;
         return categoryRepository.findById(id)
                 .map(this::convertToDTO).orElse(null);
     }
 
     @Override
     public CategoryDTO delete(Long id) {
+        if (id == null) return null;
         Optional<Category> menu = categoryRepository.findById(id);
         if (menu.isPresent()) {
             categoryRepository.deleteById(id);
